protein_localization/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
tensorflow>=2.6.0
biopython>=1.79
matplotlib>=3.4.3
seaborn>=0.11.2
joblib>=1.0.1 